[#1087] Move ArrayUtils and StringUtils to utils directory
