+++ frontend/src/services/WpApi.ts
@@ -3,7 +3,7 @@
 import * as request from 'superagent';
 
 import config from '../config';
-import {getValidVPJSON} from '../common/StringUtils';
+import {getValidVPJSON} from '../utils/StringUtils';
 
 request.parse['application/json'] = function(str: string) {
   let parsedJSON;
