+++ frontend/src/utils/ArrayUtils.ts
