+++ frontend/src/utils/StringUtils.ts
