+++ frontend/src/utils/__tests__/StringUtils-test.ts
